refactor: 三集群的绑核全局变量封装工具已经应用完毕
