+++ src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod affinity_setter;
 pub mod global_cpu_set;
+pub mod global_cpu_utils;
 pub mod node_reader;
